+++ lib/flutter_deck.dart
@@ -1,10 +1,9 @@
 /// The power of [Flutter](https://flutter.dev/), in your presentations.
 library flutter_deck;
 
+export 'src/configuration/configuration.dart';
 export 'src/flutter_deck.dart';
 export 'src/flutter_deck_app.dart';
-export 'src/flutter_deck_aspect_ratio.dart';
-export 'src/flutter_deck_configuration.dart';
 export 'src/flutter_deck_slide.dart';
 export 'src/flutter_deck_speaker_info.dart';
 export 'src/templates/split_slide.dart' show SplitSlideRatio;